PickerPanel renames and cleanup
