+++ packages/@uppy/dashboard/src/components/PickerPanelContent.js
