+++ packages/@uppy/dashboard/src/components/PickerPanelTopBar.js
